Заменил тим баланса долга, раньше был float, а теперь BigDecimal.

+++ src/FirstHomeWork/interfaces/MeasurableInRubles.java
@@ -1,8 +1,10 @@
 package FirstHomeWork.interfaces;
 
+import java.math.BigDecimal;
+
 public interface MeasurableInRubles {
 
-    float getBalance();
+    BigDecimal getBalance();
 
     default long getAnswer() {
         return 42;
